Fixed minor file naming errors and fixed login navigation bug
