+++ src/navigation/auth-stack.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { LoginScreen, AuthEmail } from 'src/screens';
+import { LoginScreen, WelcomeScreen } from 'src/screens';
 
 export type AuthStackParams = {
     Email?: undefined;
@@ -14,20 +14,17 @@ export type AuthStackParams = {
 const StackNav = createNativeStackNavigator<AuthStackParams>();
 
 export const AuthStackNavigator: React.FC<any> = () => {
-    useEffect(() => {
-        // console.log('Rendering');
-    });
 
     return (
         <StackNav.Navigator screenOptions={{ gestureEnabled: true }} initialRouteName="Email">
             <StackNav.Screen
                 name="Email"
-                component={LoginScreen}
+                component={WelcomeScreen}
                 options={{ headerShown: false, animationTypeForReplace: 'pop' }}
             />
             <StackNav.Screen
                 name="AuthEmail"
-                component={AuthEmail}
+                component={LoginScreen}
                 options={{ headerShown: false }}
             />
         </StackNav.Navigator>
